+++ src/__tests__/matrix/utility.test.js
@@ -45,6 +45,11 @@ describe('utility methods', () => {
     matrix.set(0, 0, 10);
     let called = 0;
 
+    /**
+     * @this {Matrix}
+     * @param i
+     * @param j
+     */
     function cb(i, j) {
       called++;
       expect(this).toBeInstanceOf(Matrix);
@@ -548,6 +553,44 @@ describe('utility methods', () => {
       [0, 4, 9],
     ]);
     expect(m.isSymmetric()).toBe(false);
+  });
+
+  it('isDistance', () => {
+    expect(
+      new Matrix([
+        [1, 0, 2],
+        [0, 4, 9],
+        [2, 9, 3],
+      ]).isDistance(),
+    ).toBe(false);
+    expect(
+      new Matrix([
+        [1, 0, 4],
+        [0, 4, 1],
+        [2, 9, 3],
+      ]).isDistance(),
+    ).toBe(false);
+    expect(
+      new Matrix([
+        [1, 0, 2],
+        [0, 4, 9],
+      ]).isDistance(),
+    ).toBe(false);
+
+    expect(
+      new Matrix([
+        [0, 1, 2],
+        [1, 0, 9],
+        [2, 9, 0],
+      ]).isDistance(),
+    ).toBe(true);
+    expect(
+      new Matrix([
+        [0, 1, 2],
+        [1, 1, 9],
+        [2, 9, 0],
+      ]).isDistance(),
+    ).toBe(false);
   });
 
   it('isEmpty', () => {